Complete overhall of .js file & updated user interface in .html.
A friend looked over my code and made many helpful
suggestions (thanks Peter!). The game is now almost
completely functional.

The program is now able to keep track of the round
number and display the current score. It also prints
the result of each round into a log on the page
itself rather than the console. Once round 5 has
been reached, the buttons will no longer function,
signaling the end of the game.